Fixed support for all 0-or-more epub2 metadata fields
All of the fields that I previously understood to be 0-or-1 ocurrance are actually 0-or-more. This has been changed.

Some of the epub2 metadata data structures were renamed in this change to be less noisy.

I was also able to vastly simplify the Format code and re-use a more generic function for many of the tags.

The unit tests were updated as well.