Simplified commonly-used getPkgXmlFromZip to only return the XML
The other functions are now called getPkgPathXml* and return the
tuple as they always have.

But what was happening is that 99% of the consumers of this are
going to want to just get the XML document out of a book, to then
parse for data structures like Metadata. This change makes that
more convenient.