Fixed a bug reading books where the container XML document is malformed
This is the same problem we've seen in the past with the
package document having illegal characters preceding the XML
declaration. Some code reorganization was required to fix this,
moving the XML-string-hacking functions from Codec.Epub.Parse to a
common module, Codec.Epub.Util so they can be used from Codec.Epub.IO
as well.