Added another EPUB3 dcterms date type
This one is called simply 'date'. To add it, I added a constructor Date
to the DateEvent type. This change caused problems with the existing
`newtype Date = Date String` we had so that has been changed to `newtype
DateValue = DateValue ...`

Lots of small code changes were needed to make this happen.